more corrections brought to the table

+++ subscribe.php
@@ -1,8 +1,5 @@
 <?php
 
-use Symfony\Component\Mailer\Mailer;
-use Symfony\Component\Mime\Email;
-
 session_start();
 
 // Перенаправление анонимного пользователя
@@ -11,6 +8,10 @@ if (!isset($_SESSION['user'])) {
     exit;
 }
 
+use Symfony\Component\Mailer\Exception\TransportExceptionInterface;
+use Symfony\Component\Mailer\Mailer;
+use Symfony\Component\Mime\Email;
+
 require_once 'utils.php';
 require_once 'db_config.php';
 require_once 'email_config.php';
@@ -18,56 +19,58 @@ require_once 'email_config.php';
 // параметр запроса пользователя
 $user_id = filter_input(INPUT_GET, 'user_id', FILTER_SANITIZE_NUMBER_INT);
 
+// проверка, что такой пользователь существует
+$user_exists = check_user($db_connection, $user_id ?? 0);
+
 // если параметр передан и пользователь не пытается подписаться на самого себя
-if ($user_id and $user_id !== $_SESSION['user']['id']) {
+if ($user_exists and $user_id !== $_SESSION['user']['id']) {
 
-    // проверка, что такой пользователь существует
-    $user_exists = check_user($db_connection, $user_id);
+    // проверка на уже существующую пару в подписках
+    $query = 'SELECT id
+                  FROM follower_list
+              WHERE following_user_id = ' . $_SESSION['user']['id'] . '
+              AND followed_user_id = ' . $user_id;
+    $already_subscribed = get_data_from_db($db_connection, $query);
 
-    if ($user_exists) {
-        // проверка на уже существующую пару в подписках
-        $query = 'SELECT id
-                      FROM follower_list
+    if ($already_subscribed) {
+        // запрос на отписку от пользователя
+        $query = 'DELETE FROM follower_list
                   WHERE following_user_id = ' . $_SESSION['user']['id'] . '
                   AND followed_user_id = ' . $user_id;
-        $already_subscribed = get_data_from_db($db_connection, $query);
+    } else {
+        // запрос на подписку на пользователя
+        $query = 'INSERT INTO follower_list
+                         (following_user_id, followed_user_id)
+                  VALUES (' .
+            $_SESSION['user']['id'] . ', ' . $user_id . ')';
 
-        if ($already_subscribed) {
-            // запрос на отписку от пользователя
-            $query = 'DELETE FROM follower_list
-                      WHERE following_user_id = ' . $_SESSION['user']['id'] . '
-                      AND followed_user_id = ' . $user_id;
-        } else {
-            // запрос на подписку на пользователя
-            $query = 'INSERT INTO follower_list
-                             (following_user_id, followed_user_id)
-                      VALUES (' .
-                $_SESSION['user']['id'] . ', ' . $user_id . ')';
+        // Получение данных пользователя для e-mail оповещения
+        $user_data_query = 'SELECT user_name,
+                                   user_email
+                            FROM user
+                            WHERE id = ' . $user_id;
+        $user_data = get_data_from_db($db_connection, $user_data_query, 'row');
+        $user_name = $user_data['user_name'];
+        $user_email = $user_data['user_email'];
 
-            // Получение данных пользователя для e-mail оповещения
-            $user_data_query = 'SELECT user_name,
-                                       user_email
-                                FROM user
-                                WHERE id = ' . $user_id;
-            $user_data = get_data_from_db($db_connection, $user_data_query, 'row');
-            $user_name = $user_data['user_name'];
-            $user_email = $user_data['user_email'];
+        // Формирование e-mail сообщения
+        $message = new Email();
+        $message->to($user_data['user_email']);
+        $message->from('readme_blog_noreply@list.ru');
+        $message->subject("У вас новый подписчик");
+        $message->text('Здравствуйте, ' . $user_name .
+                             '. На вас подписался новый пользователь ' . $_SESSION['user']['user_name'] .
+                             '. Вот ссылка на его профиль: readme/profile.php?user_id=' . $_SESSION['user']['id']);
 
-            // Формирование e-mail сообщения
-            $message = new Email();
-            $message->to("the_lost_number@mail.ru");
-            $message->from("the_lost_number@mail.ru");
-            $message->subject("У вас новый подписчик");
-            $message->text('Здравствуйте, ' . $user_name .
-                '. На вас подписался новый пользователь ' . $_SESSION['user']['user_name'] .
-                '. Вот ссылка на его профиль: readme/profile.php?user_id=' . $_SESSION['user']['id']);
-
-            // Отправка сообщения
-            $mailer = new Mailer($transport);
+        // Отправка сообщения
+        $mailer = new Mailer($transport);
+        try {
             $mailer->send($message);
+        } catch (TransportExceptionInterface $e) {
+            $_SESSION['errors'] = 'Cannot send message, ' . $e->getMessage();
         }
-        mysqli_query($db_connection, $query);
     }
+    mysqli_query($db_connection, $query);
 }
 
 // переадресация на прошлую страницу
